app/llmctl-rag/requirements.txt
chromadb>=0.5.0
openai>=1.0.0
tiktoken>=0.7.0
watchdog>=4.0.0
flask>=3.0.0
sqlalchemy>=2.0.0
pymupdf>=1.24.0
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.10.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
celery>=5.3.0
redis>=5.0.0
